+++ src/blob/BlobRequestListenerFactory.ts
@@ -29,6 +29,7 @@ import StrictModelMiddlewareFactory, {
 } from "./middlewares/StrictModelMiddlewareFactory";
 import IBlobMetadataStore from "./persistence/IBlobMetadataStore";
 import { DEFAULT_CONTEXT_PATH } from "./utils/constants";
+import BlobTokenAuthenticator from "./authentication/BlobTokenAuthenticator";
 
 /**
  * Default RequestListenerFactory based on express framework.
@@ -147,7 +148,8 @@ export default class BlobRequestListenerFactory
           this.accountDataStore,
           this.metadataStore,
           logger
-        )
+        ),
+        new BlobTokenAuthenticator(this.accountDataStore, logger)
       ])
     );