+++ src/blob/handlers/PageBlobHandler.ts
@@ -1,4 +1,5 @@
 import IExtentStore from "../../common/persistence/IExtentStore";
+import { convertRawHeadersToMetadata } from "../../common/utils/utils";
 import BlobStorageContext from "../context/BlobStorageContext";
 import NotImplementedError from "../errors/NotImplementedError";
 import StorageErrorFactory from "../errors/StorageErrorFactory";
@@ -96,10 +97,15 @@ export default class PageBlobHandler extends BaseHandler
     //   );
     // }
 
+    // Preserve metadata key case
+    const metadata = convertRawHeadersToMetadata(
+      blobCtx.request!.getRawHeaders()
+    );
+
     const etag = newEtag();
     const blob: BlobModel = {
       deleted: false,
-      metadata: options.metadata,
+      metadata,
       accountName,
       containerName,
       name: blobName,
@@ -114,7 +120,9 @@ export default class PageBlobHandler extends BaseHandler
         contentMD5: options.blobHTTPHeaders.blobContentMD5,
         contentDisposition: options.blobHTTPHeaders.blobContentDisposition,
         cacheControl: options.blobHTTPHeaders.blobCacheControl,
-        blobSequenceNumber: options.blobSequenceNumber,
+        blobSequenceNumber: options.blobSequenceNumber
+          ? options.blobSequenceNumber
+          : 0,
         blobType: Models.BlobType.PageBlob,
         leaseStatus: Models.LeaseStatusType.Unlocked,
         leaseState: Models.LeaseStateType.Available,
@@ -135,7 +143,8 @@ export default class PageBlobHandler extends BaseHandler
     await this.metadataStore.createBlob(
       context,
       blob,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions
     );
 
     const response: Models.PageBlobCreateResponse = {
@@ -226,7 +235,9 @@ export default class PageBlobHandler extends BaseHandler
       start,
       end,
       persistency,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions,
+      options.sequenceNumberAccessConditions
     );
 
     const response: Models.PageBlobUploadPagesResponse = {
@@ -295,7 +306,9 @@ export default class PageBlobHandler extends BaseHandler
       blob,
       start,
       end,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions,
+      options.sequenceNumberAccessConditions
     );
 
     const response: Models.PageBlobClearPagesResponse = {
@@ -329,7 +342,8 @@ export default class PageBlobHandler extends BaseHandler
       containerName,
       blobName,
       options.snapshot,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions
     );
 
     if (blob.properties.blobType !== Models.BlobType.PageBlob) {
@@ -401,7 +415,8 @@ export default class PageBlobHandler extends BaseHandler
       containerName,
       blobName,
       blobContentLength,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions
     );
 
     const response: Models.PageBlobResizeResponse = {
@@ -436,7 +451,8 @@ export default class PageBlobHandler extends BaseHandler
       blobName,
       sequenceNumberAction,
       options.blobSequenceNumber,
-      options.leaseAccessConditions
+      options.leaseAccessConditions,
+      options.modifiedAccessConditions
     );
 
     const response: Models.PageBlobUpdateSequenceNumberResponse = {