+++ ChangeLog.md
@@ -2,6 +2,15 @@
 
 > Note. This file includes changes after 3.0.0-preview. For legacy Azurite changes, please goto GitHub [releases](https://github.com/Azure/Azurite/releases).
 
+## 2020.03 Version 3.6.0
+
+- Supported conditional headers.
+- Compatible with upper case or lower case of x-ms-sequence-number-action values.
+- Fixed issue that x-ms-blob-sequence-number of 0 should be returned for HEAD requests on Page blob.
+- Uploading blocks with different lengths of IDs to the same blob will fail.
+- Check if block blob exists should fail if blocks are all uncommitted.
+- Case sensitive with metadata keys.
+
 ## 2020.02 Version 3.5.0
 
 - Bump up Azure Storage service API version to 2019-07-07.